Make a hook field available to TriggerBindings.
This adds a "hook" field for bindings to use, this is extracted by the GitHub
interceptor.

If the GitHub interceptor is used in an Trigger, then the hook will be
available, this hook is somewhat standardised across GitHub and GitLab at
least.

e.g. in a TemplateBinding the value "$(hook.Ref)" would get the ref that was
pushed to for a Push hook e.g. "refs/heads/master".